# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 103 Agency OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: List all projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with phases, milestones, tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all tasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        budget:
          type: number
          nullable: true
        dueDate:
          type: string
          nullable: true
        clientId:
          type: integer
        clientName:
          type: string
        organizationId:
          type: integer
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, status, clientId, clientName, organizationId, taskCount, completedTaskCount, createdAt, updatedAt]
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        budget:
          type: number
          nullable: true
        dueDate:
          type: string
          nullable: true
        clientId:
          type: integer
      required: [name, status, clientId]
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        budget:
          type: number
          nullable: true
        dueDate:
          type: string
          nullable: true
        clientId:
          type: integer
        clientName:
          type: string
        organizationId:
          type: integer
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        phases:
          type: array
          items:
            $ref: "#/components/schemas/Phase"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, status, clientId, clientName, organizationId, taskCount, completedTaskCount, phases, milestones, tasks, createdAt, updatedAt]
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          nullable: true
        projectId:
          type: integer
        projectName:
          type: string
        assigneeId:
          type: integer
          nullable: true
        assigneeName:
          type: string
          nullable: true
        organizationId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, priority, projectId, projectName, organizationId, createdAt, updatedAt]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          nullable: true
        projectId:
          type: integer
        assigneeId:
          type: integer
          nullable: true
      required: [title, status, priority, projectId]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          nullable: true
        assigneeId:
          type: integer
          nullable: true
...
    Phase:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
        status:
          type: string
        projectId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, order, status, projectId, createdAt, updatedAt]
...
    Milestone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        completed:
          type: boolean
        projectId:
          type: integer
        phaseId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, completed, projectId, createdAt, updatedAt]
...