# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 103 Agency OS API
 * OpenAPI spec version: 0.1.0
 */
import type { Milestone } from "./milestone";
import type { Phase } from "./phase";
import type { Task } from "./task";

export interface ProjectDetail {
  id: number;
  name: string;
  description?: string | null;
  status: string;
  budget?: number | null;
  dueDate?: string | null;
  clientId: number;
  clientName: string;
  organizationId: number;
  taskCount: number;
  completedTaskCount: number;
  phases: Phase[];
  milestones: Milestone[];
  tasks: Task[];
  createdAt: string;
  updatedAt: string;
}