# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 103 Agency OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: string;
  priority: string;
  dueDate?: string | null;
  projectId: number;
  projectName: string;
  assigneeId?: number | null;
  assigneeName?: string | null;
  organizationId: number;
  createdAt: string;
  updatedAt: string;
}